Initial commit
First commit for the AnyMacro Preprocessor which makes C-style
preprocessor macros available for any file type.

It's currently written in JavaScript due to its simplicity but
I might rewrite it in another language later.

This program is licensed under the GPLv3 license, for more
information visit the following web page:

    https://www.gnu.org/licenses/gpl-3.0.en.html

View the translation for your language on the following page:

    https://www.gnu.org/licenses/gpl.html

For the whole license in plain text, view the LICENSE file or
the NOTICE file which is a symbolic link to the LICENSE file.