Add support for import and fix undef
Now you can use the import macro to import other files:
  #import extra.txt

You can do so in any file, even in the imported files.
Importing files is as if you would paste the content of
the importing file at the exact place where the import
macro is. It is blocking, so it will first process the
imported file and then return back to the previous file.

--

Also fixed #undef

--

Some other bugfixes and improvements were done as well.